requirements.txt
# Advanced ML frameworks
transformers>=4.40.0
scikit-learn>=1.4.0
optuna>=3.6.0
lightgbm>=4.3.0
# Security and dependency scanning
bandit>=1.7.0
pip-audit>=2.4.0
# Core AI/ML Libraries
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Processing
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0

# Jupyter and Development
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Model Development and Deployment
mlflow>=2.5.0
wandb>=0.15.0
transformers>=4.30.0
datasets>=2.14.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Observability & Tracing
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-exporter-otlp>=1.24.0
# Prometheus client
prometheus-client>=0.19.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # For PostgreSQL

# Rate limiting and Redis
slowapi>=0.1.8
redis>=5.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Cloud storage and notifications
boto3>=1.28.0  # AWS S3
slack_sdk>=3.27.0

# Security
bandit>=1.7.0
